refactor[[api-server]]: 1차 검증 로직 위치 수정
[- OrderEventCheckService에서 1차 검증 수행]
